lotsa stuff, next step - delete data from db before writing to db

+++ src/main/java/lt/codeacademy/u8/tarpinisMeteo/meteo/forecast/ForecastTimestamp.java
@@ -1,17 +1,16 @@
 package lt.codeacademy.u8.tarpinisMeteo.meteo.forecast;
 
-public class ForecastTimestamp{
+public class ForecastTimestamp {
     public String forecastTimeUtc;
     public double airTemperature;
     public double feelsLikeTemperature;
-    public int windSpeed;
-    public int windGust;
-    public int windDirection;
-    public int cloudCover;
-    public int seaLevelPressure;
-    public int relativeHumidity;
+    public double windSpeed;
+    public double windGust;
+    public double windDirection;
+    public double cloudCover;
+    public double seaLevelPressure;
+    public double relativeHumidity;
     public double totalPrecipitation;
     public String conditionCode;
-
-
 }
+
